refactor: KATA-001 improving infra tests
