+++ src/infrastructure/test/persistence/temperature-capture.repository.adapter.spec.ts
@@ -1,11 +1,11 @@
 import { Repository } from 'typeorm';
-import { startTestDatabase, stopTestDatabase, getDataSource } from '../setup/test-database';
+import { startTestDatabase, stopTestDatabase } from '../setup/setup-test-database';
 import { TemperatureCaptureRepositoryAdapter } from '../../src/persistence/adapters/temperature-capture.repository.adapter';
 import { TemperatureCaptureEntity } from '../../src/persistence/entities/temperature-capture.entity';
 import { v4 as uuidv4 } from 'uuid';
 import { TemperatureState } from '../../../domain/domain-contract/models/temperature-state.enum';
 
-describe('TemperatureCaptureRepositoryAdapter', () => {
+describe('TemperatureCaptureRepositoryAdapterTest', () => {
   let adapter: TemperatureCaptureRepositoryAdapter;
   let repo: Repository<TemperatureCaptureEntity>;
 
@@ -20,7 +20,11 @@ describe('TemperatureCaptureRepositoryAdapter', () => {
 
   //region save
   it('save_shouldPersistCapture_whenValidCaptureProvided', async () => {
-    await adapter.save({ id: uuidv4(), value: 25.5, state: TemperatureState.WARM, capturedAt: new Date() });
+    await adapter.save({
+      id: uuidv4(),
+      value: 25.5,
+      state: TemperatureState.WARM, capturedAt: new Date()
+    });
 
     const results = await adapter.findLastN(10);
     expect(results).toHaveLength(1);
@@ -31,18 +35,30 @@ describe('TemperatureCaptureRepositoryAdapter', () => {
   //region findLastN
   it('findLastN_shouldReturnOrderedByDateDesc_whenMultipleCapturesExist', async () => {
     for (let i = 0; i < 5; i++) {
-      await adapter.save({ id: uuidv4(), value: 20 + i, state: TemperatureState.WARM, capturedAt: new Date(Date.now() + i * 1000) });
+      await adapter.save({
+        id: uuidv4(),
+        value: 20 + i,
+        state: TemperatureState.WARM,
+        capturedAt: new Date(Date.now() + i * 1000)
+      });
     }
 
     const results = await adapter.findLastN(3);
 
     expect(results).toHaveLength(3);
     expect(results[0].value).toBe(24);
+    expect(results[1].value).toBe(23);
+    expect(results[2].value).toBe(22);
   });
 
   it('findLastN_shouldRespectMaxCount_whenMoreCapturesThanLimit', async () => {
     for (let i = 0; i < 20; i++) {
-      await adapter.save({ id: uuidv4(), value: i, state: TemperatureState.WARM, capturedAt: new Date(Date.now() + i * 100) });
+      await adapter.save({
+        id: uuidv4(),
+        value: i,
+        state: TemperatureState.WARM,
+        capturedAt: new Date(Date.now() + i * 100)
+      });
     }
 
     const results = await adapter.findLastN(15);
